Exporting tailwind config for use in project
Very work in progress. I'm cutting back to just primary and secondary while I'm prototyping, and for now I'm just exporting the whole tailwind config. I think later there will be a more complicated system that generates what goes in here, but for now we just going forward with this